Moved common methods between check and process to helper module
